Feat/register login page impl (#55)
* feat(authentication): add login and register pages with routing

Add new Login and Register components to handle user authentication. Include routing for these pages in the router configuration. Also, add an image asset for the login page and update documentation with notes on React version conflict resolution and Tailwind CSS tips.

* feat(login): implement login page with form and social login options

Add a fully functional login page with email/password form and social login buttons for Facebook and Google. The form includes state management and input handling. The page also features a responsive design and a link to the registration page.